backend/requirements.txt
flask==2.0.1
pandas==2.2.3
numpy==2.2.4
python-dotenv==1.1.0
spacy==3.8.5
flask_cors==5.0.1
pymongo==4.8.0
Werkzeug<2.1
seaborn==0.13.2
matplotlib==3.10.1
scikit-learn==1.6.1
selenium==3.141.0
requests==2.21.0
cryptography>=3.4.7
marshmallow==3.20.1
PyYAML==6.0.1
torch>=1.9.0
tqdm==4.66.1